\name{RcmdrPlugin.WorldFlora-internal}
\title{Internal RcmdrPlugin.WorldFlora Objects}
\alias{WFO.downloadGUI}
\alias{WFO.exploreGUI}
\alias{WFO.matchGUI}
\alias{WFO.match.fuzzyjoinGUI}
\alias{WFO.oneGUI}
\alias{WFO.acceptable.matchGUI}
\alias{WFO.prepareGUI}
\alias{WFO.remember.chooseGUI}
\alias{WFO.rememberGUI}
\alias{exportExcelGUI}
\description{Internal RcmdrPlugin.WorldFlora objects.}
\details{These are not to be called by the user.}
\keyword{internal}


