\name{WorldFlora.get.started}
\alias{WorldFlora.get.started}
\alias{browseWFOmanual}
\alias{helpWorldFlora}
\alias{browseRpub}
\alias{browseRpub2023}

\title{
Getting Started with the GUI and WorldFlora
}

\description{
File with some instructions for first time users.
}

\usage{
WorldFlora.get.started()
browseWFOmanual()
helpWorldFlora()
browseRpub()
browseRpub2023()
}


\references{
World Flora Online. An Online Flora of All Known Plants.
    \url{http://www.worldfloraonline.org}

Kindt, R. 2020. WorldFlora: An R package for exact and fuzzy matching of plant names against
    the World Flora Online taxonomic backbone data.
    Applications in Plant Sciences 8(9): e11388

Kindt, R. 2023. Standardizing tree species names of GlobalTreeSearch with WorldFlora while
  testing the faster matching function of WFO.match.fuzzyjoin.
   \url{https://rpubs.com/Roeland-KINDT/996500}

}

