\name{WorldFlora.get.started}
\alias{WorldFlora.get.started}
\alias{browseWFOmanual}
\alias{helpWorldFlora}

\title{
Getting Started with the GUI and WorldFlora
}

\description{
File with some instructions for first time users.
}

\usage{
WorldFlora.get.started()
browseWFOmanual()
helpWorldFlora()
}


